Split added a new model and composable package Every composable that belonged into another major composable file was split into their respective files for imporved visualization Models once contained inside of viewmodel now contained in the model package inside of the data package
