+++ app/src/main/java/com/developersbreach/composeactors/ui/screens/actorDetails/composable/CountryInfo.kt
@@ -1,0 +1,50 @@
+package com.developersbreach.composeactors.ui.screens.actorDetails.composable
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.MaterialTheme
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ColorFilter
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.developersbreach.composeactors.R
+import com.developersbreach.composeactors.ui.animations.borderRevealAnimation
+import com.developersbreach.composeactors.utils.calculateAge
+import com.developersbreach.composeactors.utils.getPlaceOfBirth
+
+/**
+ * Second element in [ActorInfoHeader].
+ * Fetches place of birth from function [calculateAge].
+ */
+@Composable
+internal fun CountryInfo(
+    placeOfBirth: String?
+) {
+    Column(
+        modifier = Modifier.padding(16.dp),
+        horizontalAlignment = Alignment.CenterHorizontally
+    ) {
+        Box(
+            modifier = Modifier.borderRevealAnimation()
+        ) {
+            Image(
+                painterResource(id = R.drawable.ic_globe),
+                contentDescription = stringResource(R.string.cd_place_of_birth_icon),
+                colorFilter = ColorFilter.tint(color = MaterialTheme.colors.onSurface),
+                modifier = Modifier
+                    .align(alignment = Alignment.Center)
+                    .size(32.dp)
+            )
+        }
+
+        ActorInfoHeaderSubtitle(
+            subtitle = "${getPlaceOfBirth(placeOfBirth)}"
+        )
+    }
+}
