+++ app/src/main/java/com/developersbreach/composeactors/ui/screens/actorDetails/data/model/DetailsUiState.kt
@@ -1,0 +1,13 @@
+package com.developersbreach.composeactors.ui.screens.actorDetails.data.model
+
+import com.developersbreach.composeactors.data.model.ActorDetail
+import com.developersbreach.composeactors.data.model.Movie
+
+/**
+ * Models the UI state for the [ActorDetailScreen] screen.
+ */
+data class DetailsUiState(
+    val castList: List<Movie> = listOf(),
+    val actorData: ActorDetail? = null,
+    val isFetchingDetails: Boolean = false,
+)
